Revised constructor and <<

+++ timespan.h
@@ -48,9 +48,9 @@ public:
 
 private:
 
-  double hour;
-  double minute;
-  double second;
+  int hour;
+  int minute;
+  int second;
 
 };
